requirements.txt
flask
flask-cors
opencv-python
numpy
mediapipe
scikit-learn
protobuf>=4.25.0,<5.0.0
proto-plus>=1.22.3
google-generativeai
edge-tts
